pyproject.toml
[project]
name = "t2v"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "datasets>=4.8.2",
    "huggingface-hub>=1.7.1",
    "pandas>=3.0.1",
    "python-dotenv>=1.2.2",
    "ruff>=0.15.6",
    "safetensors>=0.7.0",
    "torch>=2.10.0",
]
